backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.42.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
python-multipart>=0.0.20,<1.0.0
langchain>=1.2.0,<2.0.0
langchain-community>=0.4.0,<1.0.0
langchain-google-genai>=4.2.0,<5.0.0
langchain-chroma>=1.1.0,<2.0.0
langchain-text-splitters>=1.0.0,<2.0.0
chromadb>=1.5.0,<2.0.0
pypdf>=6.0.0,<7.0.0
python-docx>=1.1.2,<2.0.0
openpyxl>=3.1.0,<4.0.0
